adding pollard p1

+++ src/main/scala/org/bu/metcs789/basics/DiscreteLog.scala
@@ -7,8 +7,8 @@ protected class DiscreteLog(modulus: Long) extends ((Long, Long) => Option[Long]
   override def apply(base: Long, value: Long): Option[Long] = {
     require(base > 0 && value > 0)
     val m = Math.ceil(Math.sqrt(modulus)).toLong
-    val powerOf: Map[Int, Long] = (0 to m.toInt).map(i => i -> FastExp(base, i).toLong % modulus).toMap
-    val C = FastExp(ModInverse(base, modulus), m).toLong // (b ^-1)^m
+    val powerOf: Map[Int, Long] = (0 to m.toInt).map(i => i -> FastExp(base, i) % modulus).toMap
+    val C = FastExp(ModInverse(base, modulus), m) // (b ^-1)^m
     var testValue = value
     for(i <- 0 to m.toInt) {
       for(j <- powerOf.keys){
